Island: Initialize the AgentPluginService
