+++ monkey/monkey_island/cc/services/__init__.py
@@ -21,5 +21,6 @@ from .agent_binary_service import (
 
 from .agent_plugin_service import IAgentPluginService, AgentPluginService
 from .agent_plugin_service import register_resources as register_agent_plugin_resources
+from .agent_plugin_service import build as build_agent_plugin_service
 
 from .log_service import setup_log_service
