+++ monkey/monkey_island/cc/services/agent_plugin_service/__init__.py
@@ -1,3 +1,4 @@
 from .i_agent_plugin_service import IAgentPluginService
 from .agent_plugin_service import AgentPluginService
 from .flask_resources import register_resources
+from .build import build
